test_consuming_strides: Keep dev array alive
The __cuda_array_interface__ doesn't hold a reference to the
object underlying it, so a reference to the object must be kept
(see discussion in Issue #4886).

This test did not keep a reference to the array it used, so if
there were sufficient pending deallocations, the array underlying
face in this test would be deleted. This commit adds a local
variable holding the underlying array in order to prevent its
premature deletion.