Replace ddt with unittest.subTest, add more transpose tests
With the removal of support for Python 2, unittest subTests are
available on all supported versions of Python. Since the use of
ddt was quite minimal (for a single test) it is removed in favor
of using subTest so that the vendored-in version of ddt can be
dropped.

Additional test cases for the transpose kernels are also added,
for other supported data types.