test_boundscheck: fix test_no_cuda_boundscheck
This test used CUDA functionality, so it needs the SerialMixin to
prevent it running in a child process after the test runner already
initialized CUDA in the parent process. It is moved into its own class
to add the SerialMixin, to preserve the ability to run other tests from
the same class to run in parallel.

It also uses a CUDA kernel without a launch configuration, which will
soon (pending PR #5061) be an error, so we add a launch configuration
to it.