fix `redirect_type_ctor`
This fixes two issues with the `redirect_type_ctor`.

a) It failed to handle >= 1 args, so the args would never arrive in the
constructor proper.
b) It didn't match potential `*args` (VarArg) and so constructing a
numba typed-list from an iterable would get stuck trying to cast the
iterable to a Numba typed list.