pin the signature
We need to pin the Python sig to be `List(iterable)` rather than
`List(*iterable)`. In order to do so, we must return a `Signature`
object with `pysig` set correctly and not just the return type. If the
`Typer` returns only the return type the signature of the `Typer` will be
used, which contains `*args`. If we change the signature of the `Typer`,
it will no longer match `List()`  AND  `List(iterable)`. Lastly, since
the `Typer` is ferried through a redirect, we need to check for the
existence of an attached `pysig` and then attach it to the redirect
object, so that the `CallableTemplate` finds it.