Fix #5408: test_set_registers_57 fails on Maxwell
The test_set_registers_* tests were a little too strict in what they
checked - the max_registers option should limit the number of registers
used, but does not guarantee that as many registers as the limit are
used. When linking for some devices, fewer than the maximum are used.

To remedy this, the tests now check that there are less than or equal to
the max registers used. An additional test is added, to ensure that the
maximum register count would otherwise have been exceeded.