Some tidying up of cudadecl
The register functions are renamed back from `intrinsic` to `register` -
the name `intrinsic` is quite overloaded in Numba already, and keeping
the names common with the attrs of the Registry class reduces the mental
overhead of reading the code.

Also delete some commented code that has been there since the
open-sourcing of NumbaPro.