Match Python/NumPy semantics in cuda.atomic.{min,max}
There appears not to be a good reason for the semantics of
cuda.atomic.min and max deviating from NumPy - the original PR which
added it (#979) makes no comment about it. Treating NaN values as
missing data also makes it more difficult to fix Issue #5458 in a manner
that is consistent between the simulator and the hardware target.

This commit changes the CUDA atomic min and max functions to match the
semantics of Python and NumPy - that is, {min,max}(a, b) always returns
a if either is a NaN. In practice, for:

    cuda.atomic.{min,max}(ary, idx, val)

the stored result is always ary[idx] when eithe of ary[idx] or val is a
NaN.