numba.typed.List: adding delitem at the compiler level
This adds the compiler wiring for `numba_list_delitem`. Also, this
changes how `pop` is implemented at the compiler level, instead of using
the c-level `numba_list_pop` it now uses a combination of c-level
`getitem` followed by a `setitem`. This fixes a segfault causes when
using `pop` to remove a refcounted type such as a NumPy array.