CUDA lib path tests: unset CUDA_PATH when CUDA_HOME unset
The library lookup mechanism in `numba.cuda.cuda_paths.get_cuda_home()`
tries `CUDA_PATH` if `CUDA_HOME` is unset. This causes a problem for the
test_library_lookup tests if `CUDA_PATH` is set, because it unsets
`CUDA_HOME` then expects not to find the CUDA toolkit by the `CUDA_PATH`
method.

The `CUDA_PATH` variable can sometimes be set on Windows, or on other
platforms where the conda-forge CuPy package is installed - it adds an
activation script to the conda environment that sets `CUDA_PATH` to
point to the conda environment.

This commit rectifies the issue by changing the tests so that they also
unset `CUDA_PATH` when `CUDA_HOME` is unset.