Return zero array from np.dot when the arguments are empty.
When np.dot(a, b) is called with empty argments, the result can be non-empty if the zero dimension is eliminated (e.g. a is m x 0 and b is 0 x n). In these cases, we should return an array of zeros. Previously, the result was initialized with np.empty, so the result contained garbage values.

Fixes #5539