CUDA: Fix #5804 test_const_record_align fails with toolkit 10.2
Now that it is easy to query the runtime version, it is also easy to fix
this test in a backward-compatible way. From 10.2 onwards, the first
three bytes are loaded as a 2-vector and a byte, whereas they used to be
loaded using a 4-vector. The test now chooses the appropriate pattern to
look for based on the runtime version.