Fix for issue5872.
Don't treat tuple as an immutable type since fields inside
of it can be mutable.  This will keep alive a modification
a field inside of a tuple to propagate outside the function.
Added ability in testing code to give a custom compare
function to be able to compare how the argument passed in to
the various versions of testing function was changed by that
function.  Also, with this change, we need separate copies
of args so that we can see those effects and so we need to
deepcopy tuples since fields inside those can change.