CUDA: Don't optimize IR before sending it to NVVM
There are various issues created by optimizing the IR before sending it
to NVVM, in particular:

- Issue #5576: Optimizing the IR before sending it to NVVM results in code
  that makes out-of-bounds accesses.
- Issue #6022: Optimization generates memset intrinsics incompatible
  with LLVM 3.4 (NVVM).
- IR auto-upgrade breaks the IR for LLVM 3.4 due to changes in atomic
  instructions. This requires a patch to the LLVM version used for
  llvmlite, which is problematic because many distributions won't carry
  this patch. (ref: https://github.com/numba/llvmlite/pull/593)

Optimizing IR prior to sending it to NVVM was originally added to work
around an NVVM bug documented as Issue #1341. The cause of issue #1341
is now resolved in NVVM, and has been since CUDA 7.5 - the test script
at https://gist.github.com/sklam/f62f1f48bb0be78f9ceb executes in a
fraction of a second and consumes around 100MB of RAM - the bug
manifested itself as a long compilation time with gigabytes of RAM
consumed.

There are some side-effects to not optimizing before sending to NVVM,
which are reflected in the tests:

- `TestCudaConstantMemory.test_const_array_3d`: Instead of reading the
  complex value as a vector of u32s, there are two individual loads of
  the real and imaginary parts as f32s. This is not expected to have a
  negative impact on performance, since coalesced reading of f32s will
  use memory bandwidth effectively.
- `TestCudaConstantMemory.test_const_record`: The generated code here
  appears less optimal than it previously was - instead of loading
  three 8-bit members in a single 32-bit read, the reads are now
  individual 8-bit reads.
- `TestCudaConstantMemory.test_const_record_align`: A similar issue to
  test_const_record.
- `TestNvvmWithoutCuda.test_nvvm_memset_fixup`: now that IR is not
  optimized prior to sending to NVVM, memsets no longer appear in the
  IR. It is desirable to keep the memset fixup functionality and test so
  that Numba and its extensions will be able to use memsets
  (e.g. from `cgutils.memset`), so this test is modified to use strings
  of IR rather than relying on memsets naturally being produced as part
  of the compilation process. It is possible to generate a memset
  without an alignment attribute on the destination with
  `cgutils.memset`, so a check for this error and a test of the check
  are also added.

For the issues with less optimal code being generated in
`TestCudaConstantMemory`, the tests for the optimal code have been moved
into new test cases that are marked as expected fails - it would be
better to keep these tests and note the sub-optimal behaviour, with the
expectation that a future version of NVVM may improve on the code it
currently generates.