CUDA: Synchronize warps in reductions
This commit adds `syncwarp()` calls to reduction kernels at appropriate
points to ensure correctness on Volta and newer GPUs.

The reduction kernels make an assumption that writes to shared memory by
threads in a warp are automatically visible to other threads in the
warp. On Volta and beyond, this is no longer true.

From §1.4.1.2 of the Volta Tuning Guide:

(https://docs.nvidia.com/cuda/volta-tuning-guide/index.html)

> Applications that assume reads and writes are implicitly visible to
> other threads in the same warp need to insert the new __syncwarp()
> warp-wide barrier synchronization instruction between steps where data
> is exchanged between threads via global or shared memory. Assumptions
> that code is executed in lockstep or that reads/writes from separate
> threads are visible across a warp without synchronization are invalid.

Under certain circumstances, the writes to shared memory across a warp
in the reduction kernel are not visible by the time they are needed.
This can be observed with the use of racecheck:

```
$ cuda-memcheck --tool racecheck python -m numba.runtests numba.cuda.tests.cudapy.test_reduction
========= CUDA-MEMCHECK
..========= WARN: Race reported between Write access at 0x00001b50 in cudapy::numba::cuda::kernels::reduction::_gpu_reduce_factory::$3clocals$3e::gpu_reduce_block_strided$241(Array<double, int=1, C, mutable, aligned>, Array<double, int=1, C, mutable, aligned>, double, bool)
=========     and Read access at 0x00001b60 in cudapy::numba::cuda::kernels::reduction::_gpu_reduce_factory::$3clocals$3e::gpu_reduce_block_strided$241(Array<double, int=1, C, mutable, aligned>, Array<double, int=1, C, mutable, aligned>, double, bool) [232 hazards]
=========

...
... much output cut ...
...

----------------------------------------------------------------------
Ran 7 tests in 3.570s

FAILED (failures=2)
========= RACECHECK SUMMARY: 14 hazards displayed (0 errors, 14 warnings)
```

With this commit, the races no longer occur, e.g.:

```
$ cuda-memcheck --tool racecheck python -m numba.runtests numba.cuda.tests.cudapy.test_reduction
========= CUDA-MEMCHECK
.......
----------------------------------------------------------------------
Ran 7 tests in 4.002s

OK
========= RACECHECK SUMMARY: 0 hazards displayed (0 errors, 0 warnings)
```

Addendum: the check in `TestReduction.test_prod_reduce` is modified to
better report the error in the output.