fix version_info if version can not be determined
The current implementation assumes that the version info can be split
into at least three parts using the period as delimiter. In certain
situations, such as running from a dowloaded release archive, or when
building from a git checkout that does not have any tags, this
assumption does not hold.

Fixes: #6125