Fix Issue #6167: Failure in test_cuda_submodules
PR #6147 broke the test_cuda_submodules test because the
test_nvvm_driver test could no longer be imported - the change from
SUPPORTED_CC to get_supported_ccs in nvvm.py was not reflected in the
CUDA simulator.

This commit fixes the issue by changing SUPPORTED_CC to
get_supported_ccs in the simulator.