Don't build with debug on non-Linux by default
Although the `-g` flag always used to be passed to the build on all
platforms, it was ignored by the Windows toolchain so it was not
actually building with debug there. Setting self.debug also attempts to
link against the Python debug library, which isn't usually available, so
this fixes an issue in the Windows build.