Fix error in cudaRuntimeGetVersion in CUDA testsuite
Removing the call to `_safe_skip` in test_warp_ops in ad46bff had the
side-effect of there being no `cuda.runtime.get_version()` call prior to
forking when running tests in parallel. This was a problem because:

- libcudart.so must be loaded before forking (or after a fork + exec,
  which the testsuite doesn't do on Linux because the default
  multiprocessing start method is fork instead of spawn, used in the
  `ParallelTestRunner`).
- Some of the compilation tests don't require a CUDA context, and
  therefore run in parallel (in a child process)
- These compilation tests use NVVM, and trigger a load of libcudart.so
  to check the runtime version via `get_supported_ccs()` in nvvm.py.

The effect was that the call to `cudaRuntimeGetVersion()` failed when
tests were run in parallel, and it appeared that there was no supported
runtime installed.

This commit rectifies the issue by calling `get_supported_ccs()` when
the tests are loaded, to force the loading of libcudart.so and
population of `_supported_ccs` in nvvm.py before the fork.