CUDA: Remove deprecated items, expect CUDA 11.1
Based on #6377 as it removes the use of deprecated options in the test
suite.

Remove items from the CUDA target that are deprecated:

- `Device.COMPUTE_CAPABILITY`: Deprecated since 2014 when NumbaPro was
  open-sourced.
- `DeviceNDArrayBase.to_host`: Likewise.
- `argtypes`, `restype`, and `bind` kwargs to the `cuda.jit` decorator:
  Deprecated in 0.51.0

There is some associated code that is also removed - e.g. that
influenced by the `bind` kwarg, and the `__writeback` member of
`DeviceNDArrayBase`.

This PR also simplifies `declare_device`, as it is only called with a
signature now.

Since CUDA 11.1 is now out, `test_runtime` is modified to expect it as a
valid version as well.