Add UUID to CUDA devices
This commit adds a `uuid` attribute to CUDA devices, which matches the
UUID given by `nvidia-smi -L`.

There are some minor grammatical fixes in the documentation, and removal
of the redundant use `(object)` in class declaration in the simulator
(which was used to declare the classes as new-style in Python 2).