Block matplotlib in test examples
The mandelbrot doc test example plots the mandelbrot set. This is nice
for the purposes of the example, but results in a plot popping up and
blocking further tests when the suite is run on a machine with
matplotlib installed.

This commit prevents plotting when running the test, by blocking the
import of matplotlib during the examples tests.