CUDA: Fix compile to PTX with debug for CUDA 11.2
The NVVM IR version metadata needs to be present in all modules passed
to NVVM. The IR version was only set when kernel wrapper functions were
generated, so device functions never had the IR version added.

This commit remedies this by adding the IR version to all modules in the
CUDA target, rather than relying on the kernel wrapper generation
function to do it.

Fixes #6719.