CUDA: Lazily add libdevice to compilation units
Modules should be lazily loaded where possible, in order to permit
certain optimizations that make compilation more efficient and improve
runtime performance.

This commit adds support for lazy loading and uses it for libdevice,
since it is never the entry point to a compilation unit. This seems to
have a noticeable effect on compilation speed. Running:

```
python -m numba.runtests numba.cuda.tests -m
```

prior to this commit gives:

```
Ran 1132 tests in 136.102s
```

on my system. With this commit, it gives:

```
Ran 1132 tests in 86.255s
```