Remove test of runtime being a supported version
This test fails if one uses a CUDA runtime version that is not
recognized as being supported by Numba. The problem with this is that a
user running the tests against a newer version of CUDA (supposing a
version 11.3, or 12.x version existed) would report this test as
failing, even though there might not be anything wrong with using the
current version of Numba with the future CUDA toolkit version. (Running
the version of Numba with a newer toolkit might not be officially
supported, but that's no reason to prevent the test suite from passing
100% with it if it otherwise would have done)

This commit removes the offending test.